+++ test/Microsoft.IdentityModel.Protocols.SignedHttpRequest.Tests/SignedHttpRequestConfigurationTestCollection.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using Xunit;
+
+namespace Microsoft.IdentityModel.Protocols.SignedHttpRequest.Tests
+{
+    [CollectionDefinition("SignedHttpRequest Configuration", DisableParallelization = true)]
+    public class SignedHttpRequestConfigurationTestCollection
+    {
+    }
+}